Correcting Clobber of main.py
